feat:security기능구현중[be03]

+++ server/src/main/java/com/preproject/stackOverflow/question/entity/Question.java
@@ -47,7 +47,7 @@ public class Question {
 
     @Column(nullable = true)
 
-    private int vote;
+    private long vote;
 
     @Column(nullable = true)
     private String tag;
